keypad setup: the wizard, against a chip that is not there
test/unit/test_keypadsetup.c: 675 checks, the last of them the one that
matters -- the whole enrolment run end to end over a socketpair standing
in for the PCF8575, and the file it writes read back by nd_keymap_load().
A keymap this wizard writes that the core cannot read is a phone with no
keys, so that assertion is the module's reason to exist.

It also gives the wizard a memory framebuffer and a real font.ttf and
checks that ink reached the panel: the wizard draws with nd_image,
nd_draw and nd_fb_update because before nd_ui_init() that is all the
drawing there is, and a silent no-op there is a blank screen at first
boot.

Two harness notes that cost an hour each:

  * a 2-byte write on an AF_UNIX stream socket is charged ~360 bytes of
    socket buffer, so the seeding is batched a pass at a time and the
    driver's own 1,088 drive words are consumed by a thread. A real chip
    consumes every transaction as it arrives; the fake one now does too.

  * -Wsign-conversion fires on `uint16 >> uint32` at -O1 and not at -O2,
    so the ASAN variant caught two shifts the default build did not.

+++ neodct/src/lib/nd_keypadsetup.c
@@ -210,7 +210,7 @@ nd_err nd_kpsetup_scan_pairs(nd_pcf8575 *chip, nd_kpsetup_pair *out, size_t max,
 
             if (bit == drive)
                 continue;
-            if (((word >> bit) & 1u) != 0u)
+            if ((((uint32_t)word >> bit) & 1u) != 0u)
                 continue; /* still pulled high: nothing shorting it here */
 
             lo = (uint8_t)(drive < bit ? drive : bit);
@@ -350,7 +350,7 @@ bool nd_kpsetup_bipartition(const nd_kpsetup_pair *pairs, size_t n, uint8_t *sid
             uint32_t nb;
 
             for (nb = 0u; nb < (uint32_t)ND_KPSETUP_MAX_PINS; nb++) {
-                if (((adj[node] >> nb) & 1u) == 0u)
+                if ((((uint32_t)adj[node] >> nb) & 1u) == 0u)
                     continue;
                 if (colour[nb] < 0) {
                     colour[nb] = (int8_t)(1 - colour[node]);
